+++ BYOJoystick/Managers/AH94RearManager.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using BYOJoystick.Controls;
 using BYOJoystick.Managers.Base;
 using VTOLVR.DLC.Rotorcraft;
@@ -25,12 +25,6 @@ namespace BYOJoystick.Managers
         {
             return GetJoysticksByPaths(name, SideJoystick, CenterJoystick);
         }
-
-        protected override void PreMapping()
-        {
-            LogInteractablesIfEnabled(ShortName);
-        }
-
         protected override void CreateFlightControls()
         {
             FlightAxisC("Joystick Pitch", "Joystick", Joysticks, CJoystick.SetPitch);