+++ BYOJoystick/Managers/T55RearManager.cs
@@ -1,4 +1,4 @@
-﻿using BYOJoystick.Controls;
+using BYOJoystick.Controls;
 using BYOJoystick.Managers.Base;
 using VTOLVR.Multiplayer;
 using MFDButtons = MFD.MFDButtons;
@@ -24,12 +24,6 @@ namespace BYOJoystick.Managers
         {
             return GetJoysticksByPaths(name, SideJoystick, CenterJoystick);
         }
-
-        protected override void PreMapping()
-        {
-            LogInteractablesIfEnabled(ShortName);
-        }
-
         protected override void CreateFlightControls()
         {
             FlightAxisC("Joystick Pitch", "Joystick (Rear)", Joysticks, CJoystick.SetPitch);